code/plot-data/requirements.txt
# Function dependencies, for example:
# package>=version
redis
google-cloud-storage==1.32.0
matplotlib==3.2.0rc1
google-cloud-error-reporting==0.30.0
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'